chore: cleaning up duplicate code
