+++ tests/test_informationmodel.py
@@ -3,11 +3,12 @@ from concepttordf import Concept
 from datacatalogtordf import Agent
 import pytest
 from rdflib import Graph, Namespace
-from rdflib.compare import graph_diff, isomorphic
+from rdflib.compare import isomorphic
 
 from modelldcatnotordf.licensedocument import LicenseDocument
 from modelldcatnotordf.modelldcatno import InformationModel
 from modelldcatnotordf.modelldcatno import ModelElement
+from tests.testutils import _dump_diff
 
 """
 A test class for testing the class InformationModel.
@@ -436,23 +437,3 @@ def test_to_graph_should_return_license_document_bnode_with_types() -> None:
         _dump_diff(g1, g2)
         pass
     assert _isomorphic
-
-
-# ---------------------------------------------------------------------- #
-# Utils for displaying debug information
-
-
-def _dump_diff(g1: Graph, g2: Graph) -> None:
-    in_both, in_first, in_second = graph_diff(g1, g2)
-    print("\nin both:")
-    _dump_turtle(in_both)
-    print("\nin first:")
-    _dump_turtle(in_first)
-    print("\nin second:")
-    _dump_turtle(in_second)
-
-
-def _dump_turtle(g: Graph) -> None:
-    for _l in g.serialize(format="turtle").splitlines():
-        if _l:
-            print(_l.decode())
